+++ src/reducer.ts
@@ -1,4 +1,22 @@
-import { Action, ActionType, State } from './types';
+import { Coordinates, FocusOptions, Marker } from './types';
+
+export enum ActionType {
+  Animate = 'ANIMATE',
+  SetFocus = 'SET_FOCUS',
+  SetActiveMarker = 'SET_ACTIVE_MANAGER',
+}
+
+export interface Action {
+  type: ActionType;
+  payload: any;
+}
+
+export interface State {
+  activeMarker?: Marker;
+  activeMarkerObject?: THREE.Object3D;
+  focus?: Coordinates;
+  focusOptions: FocusOptions;
+}
 
 export default function reducer(state: State, action: Action): State {
   const { payload, type } = action;